Move `get_run_date` from `render_report` to `utils`
We will need it in `plot` presently.
